iox-#252 added skeleton files for posh runtime
