+++ iceoryx_binding_c/CMakeLists.txt
@@ -48,7 +48,8 @@ set(${PROJECT_NAME}_DIR ${CMAKE_CURRENT_LIST_DIR}/cmake
 #
 add_library(${PROJECT_NAME}
     STATIC
-    source/subscriber.cpp
+    source/binding_c_subscriber.cpp
+    source/binding_c_posh_runtime.cpp
 )
 add_library(${PROJECT_NAMESPACE}::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
