+++ iceoryx_binding_c/include/iceoryx_binding_c/types.h
@@ -1,0 +1,36 @@
+// Copyright (c) 2020 by Robert Bosch GmbH. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef IOX_BINDING_C_TYPES_H_
+#define IOX_BINDING_C_TYPES_H_
+
+enum subscriber_SubscriptionState
+{
+    NOT_SUBSCRIBED = 0,
+    SUBSCRIBE_REQUESTED,
+    SUBSCRIBED,
+    UNSUBSCRIBE_REQUESTED,
+    WAIT_FOR_OFFER
+};
+
+enum subscriber_AllocateError
+{
+    SUCCESS,
+    TOO_MANY_CHUNKS_HELD_IN_PARALLEL,
+    NO_CHUNK_RECEIVED,
+    INTERNAL_ERROR,
+};
+
+
+#endif
